import jdk-14.0.1+1

+++ src/java.desktop/unix/native/common/java2d/x11/X11SurfaceData.h
@@ -79,7 +79,7 @@ typedef struct {
     XShmSegmentInfo     *shmSegInfo;    /* Shared Memory Segment Info */
     jint                bytesPerLine;   /* needed for ShMem lock */
     jboolean            xRequestSent;   /* true if x request is sent w/o XSync */
-    jlong               pmSize;
+    jint                pmSize;
 
     jboolean            usingShmPixmap;
     Drawable            pixmap;
